+++ src/uini.pas
@@ -39,6 +39,7 @@ begin
   ini.WriteFloat('APRS', 'latitude', Config^.Latitude);
   ini.WriteFloat('APRS', 'longitude', Config^.Longitude);
   ini.WriteInteger('APRS', 'cleanuptime', Config^.CleanupTime);
+  ini.WriteInteger('APRS', 'symbol', Config^.AprsSymbol);
   ini.WriteBool('IGATE', 'enable', Config^.IGateEnabled);
   ini.WriteString('IGATE', 'server', Config^.IGateServer);
   ini.WriteInteger('IGATE', 'port', Config^.IGatePort);
@@ -78,6 +79,7 @@ begin
   Config^.Latitude := ini.ReadFloat('APRS', 'latitude', 34.509410);
   Config^.Longitude := ini.ReadFloat('APRS', 'longitude', -32.892659);
   Config^.CleanupTime := ini.ReadInteger('APRS', 'cleanuptime', 5);
+  Config^.AprsSymbol := ini.ReadInteger('APRS', 'symbol', 9);
   Config^.IGateEnabled := ini.ReadBool('IGATE', 'enable', False);
   Config^.IGateServer := ini.ReadString('IGATE', 'server', 'rotate.aprs.net');
   Config^.IGatePort := ini.ReadInteger('IGATE', 'port', 14580);